+++ tests/unit/Runner/RunOptsTest.php
@@ -58,4 +58,26 @@ class RunOptsTest extends TestCase
         $opts->setBinaryPackage('foo');
         $this->assertSame('foo', $opts->getBinaryPackage());
     }
+
+    /**
+     * @depends testCreation
+     * @param RunOpts $opts
+     */
+    public function testSteps(RunOpts $opts)
+    {
+        self::assertNull($opts->getSteps());
+        $opts->setSteps('1,22,4');
+        self::assertIsString($opts->getSteps());
+    }
+
+    /**
+     * @depends testCreation
+     * @param RunOpts $opts
+     */
+    public function testNoManual(RunOpts $opts)
+    {
+        self::assertFalse($opts->isNoManual());
+        $opts->setNoManual(true);
+        self::assertTrue($opts->isNoManual());
+    }
 }