+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to Pipelines will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- `--no-manual` option to not stop at manual step(s). The new default is
+  to stop at steps marked manual `trigger: manual`. The first step of a
+  pipeline can not be manual, and the first step execute with `--steps`
+  will never stop even it has a `trigger: manual`.
+### Fixed
+- Base unit-test-case missing shim createConfguredMock method
+
 ## [0.0.26] - 2020-03-09
 ### Added
 - `--steps` option to specify which step(s) of a pipeline to run. Same as